Example fitted model objects used for testing msmbayes

Example fitted model objects used for testing msmbayes

Usage

infsim_model

infsim_modelc

infsim_modelp

infsim_modelpc

cav_misc

Format

An object of class msmbayes, obtained by fitting a Markov model to the dataset infsim. See data-raw/infsim.R in the source for the model specification code

infsim_modelc includes covariates on the transition intensities.

infsim_modelp is a phase-type model.

infsim_modelpc is a phase-type model with covariates.

cav_misc is the misclassification model fitted to the CAV data from msm in the "Advanced multi-state models in msmbayes" vignette. intensities.

Source

Simulated