Simulated infection testing data

Simulated infection testing data

infsim

infsim2

Format

infsim has 3600 rows, with 36 state observations for each of 100 people. Columns are:

subject Subject identifier

days Observation time (in days)

months Observation time (in moths)

state State simulated from a Markov model with no covariates

sex: "male" or "female".

age10: Age, in units of 10 years since age 50

statec: State simulated from a Markov model with covariates

statep: State simulated from a phase-type model (unused in any examples. See data-raw/infsim.R in the source for simulation settings)

statepc: State simuated from a phase-type model with covariates (unused in any examples)

A smaller dataset infsim2 has only 360 rows, from 20 people, and is simulated using a sojourn time of 60 days in the test-negative state and 10 days in test-positive.

An object of class data.frame with 360 rows and 14 columns.

Source

Simulated