Transition intensity matrix from an msmbayes model

Transition intensity matrix from an msmbayes model

Usage

qmatrix(draws, new_data = NULL, X = NULL, drop = TRUE)

Arguments

draws

Object returned by msmbayes.

new_data

Data frame with covariate values to predict for

X

Lower-level alternative to specifying new_data, for developer use only. X is a numeric matrix formed from column-binding the covariate design matrices for each transition in turn.

drop

Only used if there are no covariates supplied in new_data. Then if drop=TRUE this returns a nstates x nstates matrix, or if drop=FALSE this returns a 3D array with first dimension ncovs=1.

Value

An array or matrix of rvar objects containing the transition intensity matrix for each new prediction data point

See also

Examples

qmatrix(infsim_model)
#> rvar<1>[2,2] mean ± sd:
#>      [,1]       [,2]      
#> [1,] -0.3 ± NA   0.3 ± NA 
#> [2,]  2.1 ± NA  -2.1 ± NA 
summary(qmatrix(infsim_model))
#> # A tibble: 4 × 10
#>   variable         mean median    sd   mad     q5    q95  rhat ess_bulk ess_tail
#>   <chr>           <dbl>  <dbl> <dbl> <dbl>  <dbl>  <dbl> <dbl>    <dbl>    <dbl>
#> 1 qmatrix(infsi… -0.295 -0.295    NA     0 -0.295 -0.295    NA       NA       NA
#> 2 qmatrix(infsi…  2.14   2.14     NA     0  2.14   2.14     NA       NA       NA
#> 3 qmatrix(infsi…  0.295  0.295    NA     0  0.295  0.295    NA       NA       NA
#> 4 qmatrix(infsi… -2.14  -2.14     NA     0 -2.14  -2.14     NA       NA       NA
summary(qmatrix(infsim_model), median, ~quantile(.x, 0.025, 0.975))
#> # A tibble: 4 × 3
#>   variable                   median `2.5%`
#>   <chr>                       <dbl>  <dbl>
#> 1 qmatrix(infsim_model)[1,1] -0.295 -0.295
#> 2 qmatrix(infsim_model)[2,1]  2.14   2.14 
#> 3 qmatrix(infsim_model)[1,2]  0.295  0.295
#> 4 qmatrix(infsim_model)[2,2] -2.14  -2.14