Determine parameters of a phase-type model that approximate a parametric shape-scale distribution

Determine parameters of a phase-type model that approximate a parametric shape-scale distribution

shapescale_to_rates(
  shape,
  scale = 1,
  family = "weibull",
  canonical = FALSE,
  spline = "hermite",
  type = "vector",
  drop = TRUE
)

Arguments

shape

shape parameter. This can be vectorised.

scale

scale parameter. This can be vectorised.

family

parametric family approximated by the phase-type distribution: "weibull" or "gamma"

canonical

Return the phase-type parameters in canonical form (phase 1 sojourn rate, sojourn rate increments in subsequent states, absorption probabilities). If FALSE then phase transition rates are returned.

spline

Type of spline used to interpolate between the training points (pointwise optima) when deriving the function that best maps the shape to each phase-type parameter

drop

If shape or scale have both have one element, and drop=FALSE, a matrix with one row is returned.

list

If TRUE then separate components are returned for progression and absorption rates. Otherwise, and by default, a vector (or matrix) is returned combining all rates. If a vector is supplied for shape or scale, the returned object (or the list components) is a matrix.

Details

The approximation is determined by finding the best-approximating phase transition rates on a grid of for a given shape, repeating for a grid of shape parameters, then interpolating the results with a cubic spline. Full code is in data-raw/kl_pointwise.R in the source package. See phase5approx for the pointwise optimal parameters. The approximation is done for a scale parameter of 1, and generalised to other scales via the accelerated failure time property.

This function is not currently as efficient as it could be, since the spline interpolation is not vectorised.