Sojourn probability in a state of a msmbayes model

Sojourn probability in a state of a msmbayes model

Usage

soj_prob(draws, t, state, new_data = NULL, method = "analytic")

Arguments

draws

Object returned by msmbayes.

t

Time since state entry

state

State of interest (integer)

new_data

Data frame with covariate values to predict for

method

If "analytic" then for nphase 5 or less, an analytic solution to the matrix exponential is employed in the calculation. For nphase 6 or more, or if method="mexp" the matrix exponential is determined using numerical methods, via expm::expm().

Value

A data frame with column value giving the probability of remaining in state by time t since state entry, as an rvar object. Other columns give the time and any covariate values.

See qdf for notes on the rvar format.