Generate a dataset from the prior predictive distribution in a msmbayes model

This generates a single sample of parameters from the prior, then generates observed states from a multi-state model with those parameters. The data argument should contain the time and subject indicators at which states are to be simulated (by default), or the maximum observation time (if complete_obs=FALSE).

msmbayes_priorpred_sample(
  data,
  state = "state",
  time = "time",
  subject = "subject",
  Q,
  covariates = NULL,
  pastates = NULL,
  pafamily = "weibull",
  paspline = "linear",
  nphase = NULL,
  E = NULL,
  priors = NULL,
  complete_obs = FALSE,
  cov_format = "orig"
)

Arguments

data

Data frame giving the observed data.

state

Character string naming the observed state variable in the data. This variable must either be an integer in 1,2,...,K, where K is the number of states, or a factor with these integers as level labels. If omitted, this is assumed to be "state".

time

Character string naming the observation time variable in the data. If omitted, this is assumed to be "time".

subject

Character string naming the individual ID variable in the data. If omitted, this is assumed to be "subject".

Q

Matrix indicating the transition structure. A zero entry indicates that instantaneous transitions from (row) to (column) are disallowed. An entry of 1 (or any other positive value) indicates that the instantaneous transition is allowed. The diagonal of Q is ignored.

There is no need to "guess" initial values and put them here, as is sometimes done in msm. Initial values for fitting are determined by Stan from the prior distributions, and the specific values supplied for positive entries of Q are disregarded.

covariates

Specification of covariates on transition intensities. This should be a list of formulae. Each formula should have a left-hand side that looks like Q(r,s), and a right hand side defining the regression model for the log of the transition intensity from state \(r\) to state \(s\).

For example,

covariates = list(Q(1,2) ~ age + sex, Q(2,1) ~ age)

specifies that the log of the 1-2 transition intensity is an additive linear function of age and sex, and the log 2-1 transition intensity is a linear function of age. You do not have to list all of the intensities here if some of them are not influenced by covariates.

In models with phase-type approximated states (specified with pastates), as in standard Markov models, the numbers inside Q() refer to the observed state space. For such phase-type models, the covariate has an identical multiplicative effect on all rates of transition between phases for a given states.

In phase-type models specified with nphase, or misclassification models (specified with E), the numbers in Q() refer to transition rates on the latent state space.

pastates

This indicates which states (if any) are given a Weibull or Gamma sojourn distribution approximated by a 5-phase model. Ignored if nphase is supplied.

pafamily

"weibull" or "gamma", indicating the approximated sojourn distribution in the phased state. Either a vector of the same length as pastates, or just one to apply to all states.

paspline

"linear" or "hermite". Advanced: spline used in constructing the approximations. May remove this argument if one of these turns out to be good enough.

nphase

Only required for models with phase-type sojourn distributions specified manually (not through pastates). nphase is a vector with one element per state, giving the number of phases per state. This element is 1 for states that do not have phase-type sojourn distributions.

E

By default, msmbayes fits a (non-hidden) Markov model. If E is supplied, then a Markov model with misclassification is fitted, a type of hidden Markov model. E should then be a matrix indicating the structure of allowed misclassifications, where rows are the true states, and columns are the observed states. A zero entry in row \(r\) and column \(s\) indicates that true state \(r\) cannot be observed as state \(s\). A non-zero \((r,s)\) entry indicates that true state \(r\) may be misclassified as \(s\). The diagonal of E is ignored.

priors

A list specifying priors. Each component should be the result of a call to msmprior. Any parameters with priors not specified here are given default priors: normal with mean -2 and SD 2 for log intensities, normal with mean 0 and SD 10 for log hazard ratios, or normal(0,1) for all others (log shape, log scale and log odds parameters in phase-type approximation and misclassification models). See msmprior for more details.

If only one parameter is given a non-default prior, a single msmprior call can be supplied here instead of a list.

complete_obs

If complete_obs=FALSE (the default) intermittently-observed states are generated for the subjects and times supplied in the data argument, using msm::simmulti.msm. The returned object is a data frame made by appending these states to data.

If complete_obs=TRUE, one complete state transition history is generated using msm::sim.msm. The data argument should then consist of one row, with time giving the maximum observation time, and any covariates supplied, assumed to be time-constant. The returned object is a list.

cov_format

If "orig" the covariates are in their original form that they were supplied as. If "design" (or any other value) the covariates are returned as a design matrix, i.e. with factors converted to numeric contrasts.

Value

A data frame or a list, see msm::simmulti.msm or msm::sim.msm respectively.

Details

For phase-type approximation models, this simulates from the phase-type approximation, not the Weibull or Gamma (e.g) distribution that it is designed to approximate.