Transition intensities from an msmbayes model, presented as a tidy data frame

Transition intensities from an msmbayes model, presented as a tidy data frame

Usage

qdf(draws, new_data = NULL)

Arguments

draws

Object returned by msmbayes.

new_data

Data frame with covariate values to predict for

Value

A data frame with one row per from-state / to-state / covariate value.

Column value is in the rvar format of the posterior package, representing a sample from a posterior distribution. Use the summary function on the data frame to produce summary statistics such as the posterior median or mean (see summary.msmbayes).

See also

Examples

qdf(infsim_model)
#> # A tibble: 2 × 3
#>    from    to        value
#>   <int> <int>   <rvar[1d]>
#> 1     1     2  0.74 ± 0.34
#> 2     2     1  4.26 ± 1.98
summary(qdf(infsim_model))
#>   from to      mean    median        sd       mad        q5      q95      rhat
#> 1    1  2 0.7386393 0.6628834 0.3413701 0.2793395 0.3238488 1.393055 0.9997840
#> 2    2  1 4.2616952 3.8638645 1.9795293 1.6670112 1.8419289 7.931371 0.9997983
#>   ess_bulk ess_tail
#> 1 3828.292 4002.698
#> 2 3788.614 3964.572
summary(qdf(infsim_model), median, ~quantile(.x, c(0.025, 0.975)))
#>   from to    median      2.5%    97.5%
#> 1    1  2 0.6628834 0.2791034 1.601683
#> 2    2  1 3.8638645 1.6060798 9.027934

qdf(infsim_modelc,
    new_data = data.frame(sex=c("female","male")))
#> # A tibble: 4 × 4
#>    from    to        value sex   
#>   <int> <int>   <rvar[1d]> <chr> 
#> 1     1     2  0.70 ± 0.35 female
#> 2     1     2  0.78 ± 0.38 male  
#> 3     2     1  4.24 ± 1.94 female
#> 4     2     1  4.24 ± 1.94 male