Transition intensity matrix from an msmbayes model

Transition intensity matrix from an msmbayes model

qmatrix(draws, new_data = NULL, X = NULL, drop = TRUE)

Arguments

draws

Object returned by msmbayes.

new_data

Data frame with covariate values to predict for

X

Lower-level alternative to specifying new_data, for developer use only. X is a numeric matrix formed from column-binding the covariate design matrices for each transition in turn.

drop

Only used if there are no covariates supplied in new_data. Then if drop=TRUE this returns a nstates x nstates matrix, or if drop=FALSE this returns a 3D array with first dimension ncovs=1.

Value

An array or matrix of rvar objects containing the transition intensity matrix for each new prediction data point

See also

Examples

qmatrix(infsim_model)
#> rvar<4000>[2,2] mean ± sd:
#>      [,1]          [,2]         
#> [1,] -0.74 ± 0.34   0.74 ± 0.34 
#> [2,]  4.26 ± 1.98  -4.26 ± 1.98 
summary(qmatrix(infsim_model))
#> # A tibble: 4 × 10
#>   variable         mean median    sd   mad     q5    q95  rhat ess_bulk ess_tail
#>   <chr>           <dbl>  <dbl> <dbl> <dbl>  <dbl>  <dbl> <dbl>    <dbl>    <dbl>
#> 1 qmatrix(infsi… -0.739 -0.663 0.341 0.279 -1.39  -0.324  1.00    3829.    4003.
#> 2 qmatrix(infsi…  4.26   3.86  1.98  1.67   1.84   7.93   1.00    3790.    3965.
#> 3 qmatrix(infsi…  0.739  0.663 0.341 0.279  0.324  1.39   1.00    3829.    4003.
#> 4 qmatrix(infsi… -4.26  -3.86  1.98  1.67  -7.93  -1.84   1.00    3790.    3965.
summary(qmatrix(infsim_model), median, ~quantile(.x, c(0.025, 0.975)))
#> # A tibble: 4 × 4
#>   variable                   median `2.5%` `97.5%`
#>   <chr>                       <dbl>  <dbl>   <dbl>
#> 1 qmatrix(infsim_model)[1,1] -0.663 -1.60   -0.279
#> 2 qmatrix(infsim_model)[2,1]  3.86   1.61    9.03 
#> 3 qmatrix(infsim_model)[1,2]  0.663  0.279   1.60 
#> 4 qmatrix(infsim_model)[2,2] -3.86  -9.03   -1.61