Phase-type expansion of a transition intensity matrix to create a non-Markov multi-state model

Convert a multi-state model intensity matrix with one or more non-Markov states to an intensity matrix on a phase-type state space, where the non-Markov states are modelled with a phase-type approximation of a shape/scale distribution.

Usage

qphaseapprox(
  qmatrix,
  pastates,
  shape,
  scale,
  family = "weibull",
  spline = "linear",
  att = TRUE
)

Arguments

qmatrix

Intensity matrix on the observable state space. Only the rates for transitions out of Markov states are used, and values of rates for transitions out of the non-Markov state are ignored, unless there are competing next states. In that case the relative value of the intensities are interpreted as the transition probability to each next state. These transition probabilities are multiplied by the phase transition rates of the sojourn distribution in the non-Markov state to get the transition rates from the phases to the destination state.

shape

shape parameter

scale

scale parameter

family

parametric family approximated by the phase-type distribution: "weibull" or "gamma"

spline

Type of spline used to interpolate between the training points (pointwise optima) when deriving the function that best maps the shape to each phase-type parameter

att

keep attributes indicating progression and absorption states

TODO do we really need these, they are messy.

Value

Intensity matrix on the latent state space.