Summarise basic parameter estimates from an msmbayes model

Summarise basic parameter estimates from an msmbayes model

Usage

# S3 method for class 'msmbayes'
summary(object, log = FALSE, time = FALSE, ...)

Arguments

object

Object returned by msmbayes.

log

Present log transition intensities and log hazard ratios, rather than transition intensities and hazard ratios.

time

Present inverse transition intensities (i.e. mean times to events)

...

Further arguments passed to both qdf and loghr.

Value

A data frame with one row for each basic model parameter, plus rows for the mean sojourn times. The posterior distribution for the parameter is encoded in the column value, which has the rvar data type defined by the posterior package. This distribution can be summarised in any way by calling summary again on the data frame (see the examples).

A string summarising a sample from the prior distribution, as a median and 95% equal-tailed credible interval, is given in the prior column.

Transition intensities, or transformations of transition intensities, are those for covariate values of zero.

Remaining parameters (in non-HMMs) are log hazard ratios for covariate effects.

See also

Examples

summary(infsim_model)
#> # A tibble: 4 × 5
#>   name   from    to        value  rhat
#>   <chr> <int> <int>   <rvar[1d]> <dbl>
#> 1 q         1     2  0.74 ± 0.34  1.00
#> 2 q         2     1  4.26 ± 1.98  1.00
#> 3 mst       1    NA  1.65 ± 0.77  1.00
#> 4 mst       2    NA  0.29 ± 0.13  1.00
summary(summary(infsim_model))
#>   name from to      rhat      mean    median        sd       mad        q5
#> 1    q    1  2 0.9997840 0.7386393 0.6628834 0.3413701 0.2793395 0.3238488
#> 2    q    2  1 0.9997983 4.2616952 3.8638645 1.9795293 1.6670112 1.8419289
#> 3  mst    1 NA 0.9997810 1.6478648 1.5085609 0.7749978 0.6495299 0.7178468
#> 4  mst    2 NA 0.9998044 0.2853870 0.2588083 0.1327230 0.1116803 0.1260816
#>         q95      rhat ess_bulk ess_tail
#> 1 1.3930554 0.9997840 3828.292 4002.698
#> 2 7.9313709 0.9997983 3788.614 3964.572
#> 3 3.0878606 0.9997810 3828.292 4002.698
#> 4 0.5429091 0.9998044 3788.614 3964.572
summary(summary(infsim_model), median, ~quantile(.x, 0.025, 0.975))
#>   name from to      rhat    median      2.5%
#> 1    q    1  2 0.9997840 0.6628834 0.2791034
#> 2    q    2  1 0.9997983 3.8638645 1.6060798
#> 3  mst    1 NA 0.9997810 1.5085609 0.6243433
#> 4  mst    2 NA 0.9998044 0.2588083 0.1107673